Add EventsFrom, add model.types.ts
